patch ActiveRecord::Associations::CollectionAssociation for ids_reader.
