+++ lib/ar_lazy_preload/railtie.rb
@@ -3,6 +3,7 @@
 require "ar_lazy_preload/active_record/base"
 require "ar_lazy_preload/active_record/relation"
 require "ar_lazy_preload/active_record/association"
+require "ar_lazy_preload/active_record/collection_association"
 require "ar_lazy_preload/active_record/merger"
 require "ar_lazy_preload/active_record/association_relation"
 require "ar_lazy_preload/active_record/collection_proxy"
@@ -22,6 +23,7 @@ module ArLazyPreload
           ActiveRecord::Associations::Association
         ].each { |klass| klass.prepend(Association) }
 
+        ActiveRecord::Associations::CollectionAssociation.prepend(CollectionAssociation)
         ActiveRecord::Associations::CollectionProxy.prepend(CollectionProxy)
       end
     end
